requirements.txt
langchain ==0.2.5
langchainhub ==0.1.20
langchain-chroma ==0.1.1
langchain-community ==0.2.5
langchain_huggingface == 0.0.3
transformers ==4.41.2
bitsandbytes ==0.43.1
accelerate ==0.31.0
python-dotenv ==1.0.1
langchain-openai
chainlit
scrapy
beautifulsoup4
google-search-results
langchain-chroma